anticipation: add deleteBulk handle on didMount
